tools/alert_server/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
httpx>=0.24.0
python-dotenv>=1.0.0
pydantic>=1.10.0
# Optional (for Telegram notifications). Install only if you want notifications via Telegram
# python-telegram-bot>=20.0
